requirements.txt
langgraph==0.2.28
langchain==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.6
python-dotenv==1.0.0
openai==1.51.0
